Properly show external workspaces
We are going away with the concept of a personal workspace. Right now personal workspaces in the CLI have both projects the user is a member of and projects owned by the workspace created for the user when they signed up.

And that workspace created for the user when they signup now has a team, so it appears twice in the selector.

To solve this we stop using `me.projects` and go throught he externalWorkspaces query, displaying each external project in its workspace.