Log 413 reason
We assume it's because the file is too large, but we don't know where the limit comes from and if that's the case. A middleware we use limits multipart field size for example. Logs this so we try to narrow the cause, so we can document the upload limit.

Related to: PRO-4145