ssh-command: Allow passing one or two of `-p`, `-e` and `-s` (#679)
* ssh-command: Allow passing one or two of `-p`, `-e` and `-s`

Previously, `railway ssh` required you to either be in a linked
directory or pass all three `-p/--project`, `-e/--environment` and
`-s/--service` arguments.

Now, if only one or two arguments are passed the ones that aren't passed
will default to what was linked.

This allows more easily connecting to e.g. the production environment if
you linked to the staging one, by simply passing:

    railway ssh -e <PRODUCTION_ENV_ID>

similarly if you want to connect to a service that wasn't linked to the
current directory but is in the same project and environment you can now
do:

    railway ssh -s <SERVICE_ID>

* ssh-command: Allow passing service name on top of service ID